refactor: 세션보트팀 엔티티 도메인별 분리
