+++ app/src/main/resources/db/session-board-ddl.sql
@@ -31,7 +31,7 @@ CREATE TABLE generation (
 -- ------------------------------------------------------------
 -- EVENT_POST
 -- ------------------------------------------------------------
-CREATE TYPE post_status AS ENUM ('draft', 'published', 'hidden');
+CREATE TYPE post_status AS ENUM ('DRAFT', 'PUBLISHED', 'HIDDEN');
 
 CREATE TABLE event_post (
                             id            BIGINT      GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
@@ -41,7 +41,7 @@ CREATE TABLE event_post (
                             title         TEXT        NOT NULL,
                             body          TEXT,
                             tags          TEXT[]      NOT NULL DEFAULT '{}',
-                            status        post_status NOT NULL DEFAULT 'draft',
+                            status        post_status NOT NULL DEFAULT 'DRAFT',
                             like_count    INT         NOT NULL DEFAULT 0,
                             published_at  TIMESTAMPTZ,
                             created_at    TIMESTAMPTZ NOT NULL DEFAULT now()
@@ -50,7 +50,7 @@ CREATE TABLE event_post (
 -- ------------------------------------------------------------
 -- POST_IMAGE
 -- ------------------------------------------------------------
-CREATE TABLE post_image (
+CREATE TABLE event_post_image (
                             id         BIGINT      GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
                             post_id    BIGINT      NOT NULL REFERENCES event_post(id) ON DELETE CASCADE,
                             url        TEXT        NOT NULL,
@@ -61,7 +61,7 @@ CREATE TABLE post_image (
 -- ------------------------------------------------------------
 -- LIKE
 -- ------------------------------------------------------------
-CREATE TABLE "like" (
+CREATE TABLE event_post_like (
                         id         BIGINT      GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
                         user_id    BIGINT      NOT NULL REFERENCES users(id) ON DELETE CASCADE,
                         post_id    BIGINT      NOT NULL REFERENCES event_post(id) ON DELETE CASCADE,
@@ -72,11 +72,11 @@ CREATE TABLE "like" (
 -- ------------------------------------------------------------
 -- COMMENT
 -- ------------------------------------------------------------
-CREATE TABLE comment (
+CREATE TABLE event_post_comment (
                          id         BIGINT      GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
                          post_id    BIGINT      NOT NULL REFERENCES event_post(id) ON DELETE CASCADE,
                          author_id  BIGINT      NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-                         parent_id  BIGINT      REFERENCES comment(id) ON DELETE CASCADE,
+                         parent_id  BIGINT      REFERENCES event_post_comment (id) ON DELETE CASCADE,
                          content    TEXT        NOT NULL,
                          created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
@@ -84,14 +84,14 @@ CREATE TABLE comment (
 -- ------------------------------------------------------------
 -- SESSION
 -- ------------------------------------------------------------
-CREATE TYPE session_status AS ENUM ('scheduled', 'ongoing', 'done');
+CREATE TYPE session_status AS ENUM ('SCHEDULED', 'ONGOING', 'DONE');
 
 CREATE TABLE session (
                          id            BIGINT         GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
                          generation_id BIGINT         NOT NULL REFERENCES generation(id) ON DELETE RESTRICT,
                          week_label    TEXT,
                          title         TEXT           NOT NULL,
-                         status        session_status NOT NULL DEFAULT 'scheduled',
+                         status        session_status NOT NULL DEFAULT 'SCHEDULED',
                          started_at    TIMESTAMPTZ
 );
 
@@ -131,7 +131,7 @@ CREATE TABLE note_link (
 -- ------------------------------------------------------------
 -- RESOURCE
 -- ------------------------------------------------------------
-CREATE TYPE resource_visibility AS ENUM ('public', 'member', 'private');
+CREATE TYPE resource_visibility AS ENUM ('PUBLIC', 'MEMBER', 'PRIVATE');
 
 CREATE TABLE resource (
                           id          BIGINT              GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
@@ -141,7 +141,7 @@ CREATE TABLE resource (
                           name        TEXT                NOT NULL,
                           url         TEXT                NOT NULL,
                           size_label  TEXT,
-                          visibility  resource_visibility NOT NULL DEFAULT 'member',
+                          visibility  resource_visibility NOT NULL DEFAULT 'MEMBER',
                           uploaded_at TIMESTAMPTZ         NOT NULL DEFAULT now()
 );
 
@@ -170,12 +170,12 @@ CREATE INDEX idx_event_post_tags       ON event_post USING GIN(tags);
 CREATE INDEX idx_event_post_title      ON event_post USING GIN(title gin_trgm_ops);
 CREATE INDEX idx_event_post_body       ON event_post USING GIN(body  gin_trgm_ops);
 
-CREATE INDEX idx_post_image_post       ON post_image(post_id);
+CREATE INDEX idx_post_image_post       ON event_post_image(post_id);
 
-CREATE INDEX idx_like_post             ON "like"(post_id);
+CREATE INDEX idx_like_post             ON event_post_like (post_id);
 
-CREATE INDEX idx_comment_post          ON comment(post_id);
-CREATE INDEX idx_comment_parent        ON comment(parent_id);
+CREATE INDEX idx_comment_post          ON event_post_comment (post_id);
+CREATE INDEX idx_comment_parent        ON event_post_comment (parent_id);
 
 CREATE INDEX idx_session_generation    ON session(generation_id);
 CREATE INDEX idx_session_status        ON session(status);
