+++ common/src/main/java/com/study/common/entity/event/EventPost.java
@@ -1,0 +1,77 @@
+package com.study.common.entity.event;
+
+import com.study.common.entity.User;
+import com.study.common.entity.sessionboard.Generation;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.Table;
+import java.time.OffsetDateTime;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "event_post")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class EventPost {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "author_id", nullable = false)
+  private User author;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "generation_id", nullable = false)
+  private Generation generation;
+
+  @Column(nullable = false)
+  private String type;
+
+  @Column(nullable = false)
+  private String title;
+
+  private String body;
+
+  @Column(nullable = false, columnDefinition = "text[]")
+  private String[] tags = new String[0];
+
+  @Enumerated(EnumType.STRING)
+  @Column(nullable = false)
+  private EventPostStatus status = EventPostStatus.DRAFT;
+
+  @Column(name = "like_count", nullable = false)
+  private int likeCount = 0;
+
+  @Column(name = "published_at")
+  private OffsetDateTime publishedAt;
+
+  @Column(name = "created_at", nullable = false, updatable = false)
+  private OffsetDateTime createdAt;
+
+  @PrePersist
+  void prePersist() {
+    createdAt = OffsetDateTime.now();
+  }
+
+  public static EventPost of(User author, Generation generation, String type, String title) {
+    EventPost post = new EventPost();
+    post.author = author;
+    post.generation = generation;
+    post.type = type;
+    post.title = title;
+    return post;
+  }
+}
