+++ common/src/main/java/com/study/common/entity/event/EventPostComment.java
@@ -1,10 +1,8 @@
-package com.study.common.entity.sessionboard;
+package com.study.common.entity.event;
 
 import com.study.common.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -19,43 +17,29 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "event_post")
+@Table(name = "event_post_comment")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class EventPost {
+public class EventPostComment {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "post_id", nullable = false)
+  private EventPost post;
+
+  @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "author_id", nullable = false)
   private User author;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "generation_id", nullable = false)
-  private Generation generation;
+  @JoinColumn(name = "parent_id")
+  private EventPostComment parent;
 
   @Column(nullable = false)
-  private String type;
-
-  @Column(nullable = false)
-  private String title;
-
-  private String body;
-
-  @Column(nullable = false, columnDefinition = "text[]")
-  private String[] tags = new String[0];
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private PostStatus status = PostStatus.draft;
-
-  @Column(name = "like_count", nullable = false)
-  private int likeCount = 0;
-
-  @Column(name = "published_at")
-  private OffsetDateTime publishedAt;
+  private String content;
 
   @Column(name = "created_at", nullable = false, updatable = false)
   private OffsetDateTime createdAt;
@@ -65,12 +49,17 @@ public class EventPost {
     createdAt = OffsetDateTime.now();
   }
 
-  public static EventPost of(User author, Generation generation, String type, String title) {
-    EventPost post = new EventPost();
-    post.author = author;
-    post.generation = generation;
-    post.type = type;
-    post.title = title;
-    return post;
+  public static EventPostComment of(EventPost post, User author, String content) {
+    EventPostComment eventPostComment = new EventPostComment();
+    eventPostComment.post = post;
+    eventPostComment.author = author;
+    eventPostComment.content = content;
+    return eventPostComment;
+  }
+
+  public static EventPostComment ofReply(EventPost post, User author, EventPostComment parent, String content) {
+    EventPostComment eventPostComment = of(post, author, content);
+    eventPostComment.parent = parent;
+    return eventPostComment;
   }
 }
