+++ common/src/main/java/com/study/common/entity/event/EventPostLike.java
@@ -1,5 +1,6 @@
-package com.study.common.entity.sessionboard;
+package com.study.common.entity.event;
 
+import com.study.common.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -10,30 +11,31 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
 import java.time.OffsetDateTime;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "post_image")
+@Table(
+    name = "event_post_like",
+    uniqueConstraints = @UniqueConstraint(columnNames = {"user_id", "post_id"}))
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PostImage {
+public class EventPostLike {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "user_id", nullable = false)
+  private User user;
+
+  @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "post_id", nullable = false)
   private EventPost post;
-
-  @Column(nullable = false)
-  private String url;
-
-  @Column(name = "\"order\"", nullable = false)
-  private int order = 0;
 
   @Column(name = "created_at", nullable = false, updatable = false)
   private OffsetDateTime createdAt;
@@ -43,10 +45,10 @@ public class PostImage {
     createdAt = OffsetDateTime.now();
   }
 
-  public static PostImage of(EventPost post, String url) {
-    PostImage image = new PostImage();
-    image.post = post;
-    image.url = url;
-    return image;
+  public static EventPostLike of(User user, EventPost post) {
+    EventPostLike eventPostLike = new EventPostLike();
+    eventPostLike.user = user;
+    eventPostLike.post = post;
+    return eventPostLike;
   }
 }
