+++ common/src/main/java/com/study/common/entity/event/EventPostStatus.java
@@ -1,0 +1,7 @@
+package com.study.common.entity.event;
+
+public enum EventPostStatus {
+  DRAFT,
+  PUBLISHED,
+  HIDDEN
+}
