+++ common/src/main/java/com/study/common/entity/session/ResourceVisibility.java
@@ -1,4 +1,4 @@
-package com.study.common.entity.sessionboard;
+package com.study.common.entity.session;
 
 public enum ResourceVisibility {
   PUBLIC,
