+++ common/src/main/java/com/study/common/entity/session/SessionSpeaker.java
@@ -1,4 +1,4 @@
-package com.study.common.entity.sessionboard;
+package com.study.common.entity.session;
 
 import com.study.common.entity.User;
 import jakarta.persistence.Column;
@@ -9,18 +9,19 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.PrePersist;
 import jakarta.persistence.Table;
-import java.time.OffsetDateTime;
+import jakarta.persistence.UniqueConstraint;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "session_note")
+@Table(
+    name = "session_speaker",
+    uniqueConstraints = @UniqueConstraint(columnNames = {"session_id", "user_id"}))
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SessionNote {
+public class SessionSpeaker {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -31,23 +32,16 @@ public class SessionNote {
   private Session session;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "author_id", nullable = false)
-  private User author;
+  @JoinColumn(name = "user_id", nullable = false)
+  private User user;
 
-  private String body;
+  @Column
+  private String role;
 
-  @Column(name = "created_at", nullable = false, updatable = false)
-  private OffsetDateTime createdAt;
-
-  @PrePersist
-  void prePersist() {
-    createdAt = OffsetDateTime.now();
-  }
-
-  public static SessionNote of(Session session, User author) {
-    SessionNote note = new SessionNote();
-    note.session = session;
-    note.author = author;
-    return note;
+  public static SessionSpeaker of(Session session, User user) {
+    SessionSpeaker speaker = new SessionSpeaker();
+    speaker.session = session;
+    speaker.user = user;
+    return speaker;
   }
 }
