+++ common/src/main/java/com/study/common/entity/session/SessionStatus.java
@@ -1,0 +1,7 @@
+package com.study.common.entity.session;
+
+public enum SessionStatus {
+  SCHEDULED,
+  ONGOING,
+  DONE
+}
