--- common/src/main/java/com/study/common/entity/sessionboard/Comment.java
@@ -1,66 +1,0 @@
-package com.study.common.entity.sessionboard;
-
-import com.study.common.entity.User;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-
-@Entity
-@Table(name = "comments")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(name = "post_id", nullable = false)
-  private Long postId;
-
-  @Column(name = "user_id", columnDefinition = "uuid", nullable = false)
-  private UUID userId;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "parent_id")
-  @OnDelete(action = OnDeleteAction.CASCADE)
-  private Comment parent;
-
-  @Column(columnDefinition = "TEXT", nullable = false)
-  private String content;
-
-  @CreationTimestamp
-  @Column(name = "created_at", nullable = false, updatable = false)
-  private LocalDateTime createdAt;
-
-  @Builder
-  public Comment(Long postId, UUID userId, Comment parent, String content) {
-    this.postId = postId;
-    this.userId = userId;
-    this.parent = parent;
-    this.content = content;
-  }
-
-  public Long getParentId() {
-    return parent != null ? parent.getId() : null;
-  }
-
-  public void updateContent(String content) {
-    this.content = content;
-  }
-}
