--- common/src/main/java/com/study/common/entity/sessionboard/SessionStatus.java
@@ -1,7 +1,0 @@
-package com.study.common.entity.sessionboard;
-
-public enum SessionStatus {
-  scheduled,
-  ongoing,
-  done
-}
